# aspnetcore/src/IO.Swagger/Models/UserRole.cs
/*
 * connection-proposer-api
 *
 * Connection proposer API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
        /// <summary>
        /// role of the user
        /// </summary>
        /// <value>role of the user</value>
        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum UserRole
        {
            /// <summary>
            /// Enum EngineerEnum for engineer
            /// </summary>
            [EnumMember(Value = "engineer")]
            EngineerEnum = 0,
            /// <summary>
            /// Enum LeadEngineerEnum for lead-engineer
            /// </summary>
            [EnumMember(Value = "lead-engineer")]
            LeadEngineerEnum = 1,
            /// <summary>
            /// Enum IsAdministratorEnum for is-administrator
            /// </summary>
            [EnumMember(Value = "is-administrator")]
            IsAdministratorEnum = 2        }
}